Crate frame_benchmarking_cli

Contains the root BenchmarkCmd command and exports its sub-commands.

Re-exports

pub use machine::SUBSTRATE_REFERENCE_HARDWARE;

pub use machine::SUBSTRATE_REFERENCE_HARDWARE;

Structs

BasePath

The base path that is used for everything that needs to be written on disk to run a node.

BlockCmd

Benchmark the execution time of historic blocks.

ExtrinsicCmd

Benchmark the execution time of different extrinsics.

ExtrinsicFactory

Helper to manage ExtrinsicBuilder instances.

MachineCmd

Command to benchmark the hardware.

OverheadCmd

Benchmark the execution overhead per-block and per-extrinsic.

PalletCmd

Benchmark the extrinsic weight of FRAME Pallets.

SUBSTRATE_REFERENCE_HARDWARE

The hardware requirements as measured on reference hardware.

StorageCmd

Benchmark the storage speed of a chain snapshot.

Enums

BenchmarkCmd

The root benchmarking command.

Traits

ExtrinsicBuilder

Used by the benchmark to build signed extrinsics.