Struct frame_benchmarking_cli::ExtrinsicCmd

pub struct ExtrinsicCmd {
    pub shared_params: SharedParams,
    pub import_params: ImportParams,
    pub params: ExtrinsicParams,
}

Benchmark the execution time of different extrinsics.

This is calculated by filling a block with a specific extrinsic and executing the block. The result time is then divided by the number of extrinsics in that block.

NOTE: The BlockExecutionWeight is ignored in this case since it

Fields
shared_params: SharedParamsimport_params: ImportParamsparams: ExtrinsicParams
Implementations

impl ExtrinsicCmd

pub fn run<Block, BA, C>( &self, client: Arc<C>, inherent_data: InherentData, digest_items: Vec<DigestItem>, ext_factory: &ExtrinsicFactory ) -> Result<()>where Block: BlockT<Extrinsic = OpaqueExtrinsic>, BA: ClientBackend<Block>, C: BlockBuilderProvider<BA, Block, C> + ProvideRuntimeApi<Block> + HeaderBackend<Block>, C::Api: ApiExt<Block> + BlockBuilderApi<Block>,

Benchmark the execution time of a specific type of extrinsic.

The output will be printed to console.