Struct frame_benchmarking_cli::PalletCmd

pub struct PalletCmd {

    pub pallet: Option<String>,
    pub extrinsic: Option<String>,
    pub steps: u32,
    pub lowest_range_values: Vec<u32>,
    pub highest_range_values: Vec<u32>,
    pub repeat: u32,
    pub external_repeat: u32,
    pub json_output: bool,
    pub json_file: Option<PathBuf>,
    pub no_median_slopes: bool,
    pub no_min_squares: bool,
    pub output: Option<PathBuf>,
    pub header: Option<PathBuf>,
    pub template: Option<PathBuf>,
    pub hostinfo_params: HostInfoParams,
    pub output_analysis: Option<String>,
    pub output_pov_analysis: Option<String>,
    pub default_pov_mode: PovEstimationMode,
    pub heap_pages: Option<u64>,
    pub no_verify: bool,
    pub extra: bool,
    pub shared_params: SharedParams,
    pub wasm_method: WasmExecutionMethod,
    pub wasmtime_instantiation_strategy: WasmtimeInstantiationStrategy,
    pub execution: Option<String>,
    pub database_cache_size: u32,
    pub list: bool,
    pub no_storage_info: bool,
    pub worst_case_map_values: u32,
    pub additional_trie_layers: u8,
    pub json_input: Option<PathBuf>,
    pub unsafe_overwrite_results: bool,

}

Benchmark the extrinsic weight of FRAME Pallets.

Fields
pallet: Option<String>

Select a FRAME Pallet to benchmark, or * for all (in which case extrinsic must be *).

extrinsic: Option<String>

Select an extrinsic inside the pallet to benchmark, or * for all.

steps: u32

Select how many samples we should take across the variable components.

lowest_range_values: Vec<u32>

Indicates lowest values for each of the component ranges.

highest_range_values: Vec<u32>

Indicates highest values for each of the component ranges.

repeat: u32

Select how many repetitions of this benchmark should run from within the wasm.

external_repeat: u32

Select how many repetitions of this benchmark should run from the client.

NOTE: Using this alone may give slower results, but will afford you maximum Wasm memory.

json_output: bool

Print the raw results in JSON format.

json_file: Option<PathBuf>

Write the raw results in JSON format into the given file.

no_median_slopes: bool

Don’t print the median-slopes linear regression analysis.

no_min_squares: bool

Don’t print the min-squares linear regression analysis.

output: Option<PathBuf>

Output the benchmarks to a Rust file at the given path.

header: Option<PathBuf>

Add a header file to your outputted benchmarks.

template: Option<PathBuf>

Path to Handlebars template file used for outputting benchmark results. (Optional)

hostinfo_params: HostInfoParamsoutput_analysis: Option<String>

Which analysis function to use when outputting benchmarks:

min-squares (default)

median-slopes

max (max of min squares and median slopes for each value)

output_pov_analysis: Option<String>

Which analysis function to use when analyzing measured proof sizes.

default_pov_mode: PovEstimationMode

The PoV estimation mode of a benchmark if no pov_mode attribute is present.

heap_pages: Option<u64>

Set the heap pages while running benchmarks. If not set, the default value from the client is used.

no_verify: bool

Disable verification logic when running benchmarks.

extra: bool

Display and run extra benchmarks that would otherwise not be needed for weight construction.

shared_params: SharedParamswasm_method: WasmExecutionMethod

Method for executing Wasm runtime code.

wasmtime_instantiation_strategy: WasmtimeInstantiationStrategy

The WASM instantiation method to use.

Only has an effect when wasm-execution is set to compiled.

execution: Option<String>

DEPRECATED: This argument has no effect.

database_cache_size: u32

Limit the memory the database cache can use.

list: bool

List the benchmarks that match your query rather than running them.

When nothing is provided, we list all benchmarks.

no_storage_info: bool

If enabled, the storage info is not displayed in the output next to the analysis.

This is independent of the storage info appearing in the output file. Use a Handlebar template for that purpose.

worst_case_map_values: u32

The assumed default maximum size of any StorageMap.

When the maximum size of a map is not defined by the runtime developer, this value is used as a worst case scenario. It will affect the calculated worst case PoV size for accessing a value in a map, since the PoV will need to include the trie nodes down to the underlying value.

additional_trie_layers: u8

Adjust the PoV estimation by adding additional trie layers to it.

This should be set to log16(n) where n is the number of top-level storage items in the runtime, eg. StorageMaps and StorageValues. A value of 2 to 3 is usually sufficient. Each layer will result in an additional 495 bytes PoV per distinct top-level access. Therefore multiple StorageMap accesses only suffer from this increase once. The exact number of storage items depends on the runtime and the deployed pallets.

json_input: Option<PathBuf>

A path to a .json file with existing benchmark results generated with --json or --json-file. When specified the benchmarks are not actually executed, and the data for the analysis is read from this file.

unsafe_overwrite_results: bool

Allow overwriting a single file with multiple results.

This exists only to restore legacy behaviour. It should never actually be needed.

Implementations

impl PalletCmd

pub fn run<Hasher, ExtraHostFunctions>( &self, config: Configuration ) -> Result<()>where Hasher: Hash, ExtraHostFunctions: HostFunctions,

Runs the command and benchmarks a pallet.