Struct frame_benchmarking_cli::BenchmarkCmd

pub struct BenchmarkCmd {
    pub pallet: String,
    pub extrinsic: String,
    pub steps: Vec<u32>,
    pub lowest_range_values: Vec<u32>,
    pub highest_range_values: Vec<u32>,
    pub repeat: u32,
    pub raw_data: bool,
    pub no_median_slopes: bool,
    pub no_min_squares: bool,
    pub output: bool,
    pub weight_trait: bool,
    pub heap_pages: Option<u64>,
    pub no_verify: bool,
    pub extra: bool,
    pub shared_params: SharedParams,
    pub execution: Option<ExecutionStrategy>,
    pub wasm_method: WasmExecutionMethod,
    pub database_cache_size: u32,
}

The benchmark command used to benchmark FRAME Pallets.

Fields
pallet: String

Select a FRAME Pallet to benchmark, or * for all (in which case extrinsic must be *).

extrinsic: String

Select an extrinsic inside the pallet to benchmark, or * for all.

steps: Vec<u32>

Select how many samples we should take across the variable components.

lowest_range_values: Vec<u32>

Indicates lowest values for each of the component ranges.

highest_range_values: Vec<u32>

Indicates highest values for each of the component ranges.

repeat: u32

Select how many repetitions of this benchmark should run.

raw_data: bool

Print the raw results.

no_median_slopes: bool

Don't print the median-slopes linear regression analysis.

no_min_squares: bool

Don't print the min-squares linear regression analysis.

output: bool

Output the benchmarks to a Rust file.

weight_trait: bool

Output the trait definition to a Rust file.

heap_pages: Option<u64>

Set the heap pages while running benchmarks.

no_verify: bool

Disable verification logic when running benchmarks.

extra: bool

Display and run extra benchmarks that would otherwise not be needed for weight construction.

shared_params: SharedParamsexecution: Option<ExecutionStrategy>

The execution strategy that should be used for benchmarks

wasm_method: WasmExecutionMethod

Method for executing Wasm runtime code.

database_cache_size: u32

Limit the memory the database cache can use.

Implementations

impl BenchmarkCmd

pub fn run<BB, ExecDispatch>(&self, config: Configuration) -> Result<()> where
BB: BlockT + Debug,
<<<BB as BlockT>::Header as HeaderT>::Number as FromStr>::Err: Debug,
<BB as BlockT>::Hash: FromStr,
ExecDispatch: NativeExecutionDispatch + 'static,

Runs the command and benchmarks the chain.